[OU-FIX] website_sale_attribute_filter_price: Execute it in post
"website_sale.filter_products_price" view is still not available
before loading the new module specs.
