+++ openupgrade_scripts/scripts/website_sale/15.0.1.1/post-migration.py
@@ -16,6 +16,26 @@ def set_visibility_product_attribute(env):
     )
 
 
+def enable_price_filter_view(env):
+    """If you had website_sale_attribute_filter_price module installed in previous
+    version, replace it by the new core price filter.
+    """
+    for website in env["website"].search([]):
+        view = env["ir.ui.view"].search(
+            [
+                ("key", "=", "website_sale_attribute_filter_price.pricefilter"),
+                ("website_id", "=", website.id),
+            ]
+        )
+        if view:
+            # It's important to set the context to create the new view related to the
+            # current website. See write method of ir.ui.view in website module
+            website.with_context(website_id=website.id).viewref(
+                "website_sale.filter_products_price"
+            ).active = True
+            view.unlink()
+
+
 @openupgrade.migrate()
 def migrate(env, version):
     # Load noupdate changes
@@ -32,3 +52,4 @@ def migrate(env, version):
         ],
     )
     set_visibility_product_attribute(env)
+    enable_price_filter_view(env)
